rearrange emoji data and adjust imports
